app2/requirements.txt
attrs~=22.1.0
aws-cdk-lib==2.102.0
botocore==1.31.17
cattrs~=22.2.0
cdk-ecr-deployment~=2.5.30
cdk-nag~=2.27.167
constructs>=10.0.0,<11.0.0
exceptiongroup~=1.1.3
importlib-resources~=6.1.0
isort~=5.12.0
jmespath~=1.0.1
Jinja2==3.1.2
jsii==1.90.0
pluginlib~=0.9.1
publication~=0.0.3
python-dateutil~=2.8.2
PyYAML==5.3.1
s3transfer~=0.6.2
six~=1.16.0
strongtyping~=3.10.7
typeguard~=2.13.3
typing_extensions~=4.8.0
urllib3~=1.26.15
requests~=2.30.0
retry~=0.9.2
